// src/lib/types/tomorrow-io.ts
export type WeatherForecastErrorResponse = {
  code: 429001 | number;
  type: "Too Many Calls" | string;
  message: string;
};
// ...
export type WeatherForecastNow = {
  time: Date;
  cloudBase: number;
  cloudCeiling: number;
  cloudCover: number;
  dewPoint: number;
  freezingRainIntensity: number;
  humidity: number;
  precipitationProbability: number;
  pressureSurfaceLevel: number;
  rainIntensity: number;
  sleetIntensity: number;
  snowIntensity: number;
  temperature: number;
  temperatureApparent: number;
  uvHealthConcern: number;
  uvIndex: number;
  visibility: number;
  weatherCode: number;
  windDirection: number;
  windDirectionCardinal: string;
  windGust: number;
  windSpeed: number;
};